feat: 修正插件测试依赖边界 --story=136129662
