+++ src/plugins/aidev_bkplugin/pyproject.toml
@@ -17,6 +17,7 @@ opentelemetry = [
 
 [dependency-groups]
 dev = [
+    "aidev-agent[opentelemetry]>=2.2.2rc28",
     "pytest>=7.4.4,<9",
     "pytest-cov>=4.1.0,<4.2.0",
     "pytest-mock>=3.14.0",
