+++ src/plugins/aidev_bkplugin/tests/database_events/test_tracing.py
@@ -1,8 +1,5 @@
-import asyncio
-
 import pytest
 from aidev_agent.utils import tracing
-from aidev_bkplugin.models import EventDelivery
 from opentelemetry.sdk.trace import TracerProvider
 from opentelemetry.sdk.trace.export import SimpleSpanProcessor
 from opentelemetry.sdk.trace.export.in_memory_span_exporter import InMemorySpanExporter
@@ -16,30 +13,6 @@ def spans(monkeypatch):
     monkeypatch.setattr(tracing, "_agent_tracer", provider.get_tracer("test"))
     yield exporter
     provider.shutdown()
-
-
-def test_publish_claim_consume_share_entry_trace(wx_delivery_case, spans):
-    case = wx_delivery_case
-    with tracing.recording_span("web.entry") as entry:
-        case.event.value["traceContext"] = tracing.trace_headers()
-    original = dict(case.event.value["traceContext"])
-    with tracing.recording_span("unrelated-worker"):
-        case.bus.publish(case.event)
-        case.bus.publish(case.event)  # tracing must not break idempotent publishing
-        asyncio.run(case.consumer.consume_once())
-    recorded = [
-        s
-        for s in spans.get_finished_spans()
-        if s.name in ("database_event.publish", "database_event.claim", "wxbot.event.consume")
-    ]
-    assert len(recorded) == 4
-    assert all(s.context.trace_id == entry.get_span_context().trace_id for s in recorded)
-    claim = next(s for s in recorded if s.name == "database_event.claim")
-    assert claim.attributes["messaging.message.age_ms"] >= 0
-    assert claim.attributes["messaging.receive.lookup.duration_ms"] >= 0
-    assert claim.attributes["messaging.delivery.attempt"] == 1
-    delivery = EventDelivery.objects.get()
-    assert delivery.status == "delivered" and delivery.envelope["value"]["traceContext"] == original
 
 
 def test_empty_poll_does_not_create_unbounded_spans(event_case, spans):
