+++ src/plugins/aidev_wxbot/Makefile
@@ -1,7 +1,7 @@
 # 导入子目录中的 Makefile 文件
 REPO_ROOT?=$(shell git rev-parse --show-toplevel)
 ROOT_DIR?=${REPO_ROOT}/src/plugins/aidev_wxbot
-pytest = PYTHONPATH=${REPO_ROOT}/src/agent:${REPO_ROOT}/src/plugins/aidev_bkplugin:${ROOT_DIR}:$${PYTHONPATH} uv run --no-sync pytest --ds=aidev_wxbot.settings -c ${ROOT_DIR}/pyproject.toml
+pytest = PYTHONPATH=${REPO_ROOT}/src/agent:${REPO_ROOT}/src/plugins/aidev_bkplugin:${ROOT_DIR}:$${PYTHONPATH} uv run --no-sync pytest --ds=aidev_wxbot_test_settings -c ${ROOT_DIR}/pyproject.toml
 
 
 .PHONY: ALL
