+++ src/plugins/aidev_wxbot/pyproject.toml
@@ -34,7 +34,7 @@ dev = [
     "opentelemetry-sdk>=1.20",
     "pytest>=7.4.4,<9",
     "pytest-asyncio>=0.23.3,<1",
-    "pytest-django>=4.7.0,<5",
+    "pytest-django>=4.7.0,<4.13",
     "pytest-dotenv>=0.5.2,<1",
 ]
 
@@ -77,7 +77,7 @@ section-order = [
 relative-imports-order = "closest-to-furthest"
 
 [tool.pytest.ini_options]
-DJANGO_SETTINGS_MODULE = "aidev_wxbot.settings"
+DJANGO_SETTINGS_MODULE = "aidev_wxbot_test_settings"
 env_override_existing_values = 1
 env_files = [".env"]
 asyncio_default_fixture_loop_scope = "session"
