+++ src/plugins/aidev_wxbot/tests/conftest.py
@@ -3,7 +3,6 @@
 
 import sys
 from types import ModuleType
-from unittest.mock import MagicMock
 
 import pytest
 
@@ -19,14 +18,9 @@ decorators_module.inject_user_token = inject_user_token
 kit_module.decorators = decorators_module
 framework_module.kit = kit_module
 
-models_module = ModuleType("aidev_bkplugin.models")
-models_module.Checkpoint = MagicMock()
-models_module.Write = MagicMock()
-
 sys.modules.setdefault("bk_plugin_framework", framework_module)
 sys.modules.setdefault("bk_plugin_framework.kit", kit_module)
 sys.modules.setdefault("bk_plugin_framework.kit.decorators", decorators_module)
-sys.modules.setdefault("aidev_bkplugin.models", models_module)
 
 
 @pytest.fixture
