+++ src/plugins/aidev_wxbot/tests/database_events/__init__.py
@@ -1,0 +1,1 @@
+"""Database-event integration tests owned by the downstream wxbot package."""
